double counting fix
Hopefully this will avoid double counting, so now hong kong and macau data will show as separate countries, but will not be included in the tallies for Mainland China.

Am also pulling out Hong Kong to the default view, as these numbers were previously included in China and contributing to its upwards drift.