zoom bug fix
Currently there's a bug where if the user selects a custom zoom/range and then switches a UI property (selectedData, selectedRegion, or selectedScale), the graph holds on to their range and doesn't rescale.
To fix this, we watch for changes to the UI state, and if they occur, we ignore the user set range.

Finally we're making the dependence of range on selected scale in the layout more explicit.