Changed IDs to classes
Only left #root (needed as vue.js hoook as it seems) and #hide (propbably would work as class as well but for hiding and showing IDs are more save never the less) as IDs.

Only question is #title necessary at all? never been used in index.html nor in vue-definitions.js. But left it in still