Remove article and use main instead
Yuck. One thing i forgot it isn’t allowed to have main as parent containing an aside. So it wasnt possible to use main for the main content while using article for the graph and aside for the controls. using not a wrapping div with the newly introduced class ctapp containing main on previous left-column  and aside as the right column.